fix: rename `GrailsPluginDiscoverySpec` to `PluginDiscoverySpec`
And some cleanup
